feat: 유저를 업데이트 할 때 프로필 이미지를 업로드 하지 않았다면 프로필 사진을 삭제하는 것이 아닌 업데이트하지 않도록 변경 및 필요없는 null체크 제거

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -108,7 +108,10 @@ public class BasicUserService implements UserService {
                     return binaryContentRepository.save(binaryContent).getId();
                 })
                 .orElse(null);
-        binaryContentRepository.deleteById(user.getProfileId());
+
+        if(profileId != null) {
+            binaryContentRepository.deleteById(user.getProfileId());
+        }
         user.update(username, email, password, profileId);
 
         return userRepository.save(user);
